requirements.txt
absl-py                   2.3.0
accelerate                1.7.0
addict                    2.4.0
aiofiles                  24.1.0
aiohappyeyeballs          2.6.1
aiohttp                   3.12.9
aiosignal                 1.3.2
albumentations            1.3.0
altair                    5.5.0
annotated-types           0.7.0
antlr4-python3-runtime    4.9.3
anyio                     4.9.0
asttokens                 3.0.0
async-timeout             5.0.1
attrs                     25.3.0
av                        14.4.0
beautifulsoup4            4.12.3
blinker                   1.9.0
braceexpand               0.1.7
certifi                   2025.4.26
charset-normalizer        3.3.2
click                     8.2.1
cmake                     4.0.2
contourpy                 1.3.2
cycler                    0.12.1
decorator                 5.2.1
deepspeed                 0.14.0
defusedxml                0.7.1
diffusers                 0.20.2
distro                    1.9.0
easydict                  1.13
einops                    0.6.1
einops-exts               0.0.4
exceptiongroup            1.3.0
executing                 2.2.0
fastapi                   0.115.12
ffmpy                     0.3.2
filelock                  3.13.4
fire                      0.4.0
flake8                    7.0.0
Flask                     3.0.3
fonttools                 4.49.0
frozenlist                1.4.1
fschat                    0.2.26
fsspec                    2023.10.0
ftfy                      6.1.3
gradio                    3.35
gradio_client             1.10.2
groovy                    0.1.2
groundingdino             0.1.0       
grpcio                    1.72.1
h11                       0.16.0
hf-xet                    1.1.3
hjson                     3.1.0
httpcore                  1.0.9
httpx                     0.27.2
huggingface-hub           0.25.2
hydra-core                1.3.2
idna                      3.10
imageio                   2.9.0
imageio-ffmpeg            0.5.1
importlib_metadata        8.7.0
invisible-watermark       0.2.0
ipython                   8.37.0
itsdangerous              2.2.0
jedi                      0.19.2
Jinja2                    3.1.6
jiter                     0.10.0
joblib                    1.5.1
jsonschema                4.24.0
jsonschema-specifications 2025.4.1
kiwisolver                1.4.8
kornia                    0.4.1
kornia_rs                 0.1.9
latex2mathml              3.78.0
lazy_loader               0.4
lightning-utilities       0.14.3
linkify-it-py             2.0.3
lit                       18.1.8
Markdown                  3.8
markdown-it-py            2.2.0
markdown2                 2.5.3
MarkupSafe                3.0.2
matplotlib                3.10.3
matplotlib-inline         0.1.7
mccabe                    0.7.0
mdit-py-plugins           0.3.3
mdurl                     0.1.2
mpmath                    1.3.0
multidict                 6.4.4
narwhals                  1.42.0
networkx                  3.4.2
nh3                       0.2.21
ninja                     1.11.1.4
numpy                     1.26.4
nvidia-cublas-cu11        11.11.3.6
nvidia-cuda-cupti-cu11    11.8.87
nvidia-cuda-nvrtc-cu11    11.8.89
nvidia-cuda-runtime-cu11  11.8.89
nvidia-cudnn-cu11         8.7.0.84
nvidia-cufft-cu11         10.9.0.58
nvidia-curand-cu11        10.3.0.86
nvidia-cusolver-cu11      11.4.1.48
nvidia-cusparse-cu11      11.7.5.86
nvidia-ml-py              12.575.51
nvidia-nccl-cu11          2.19.3
nvidia-nvtx-cu11          11.8.86
omegaconf                 2.3.0
open-clip-torch           2.24.0
openai                    1.55.3
opencv-python             4.9.0.80
opencv-python-headless    4.11.0.86
orjson                    3.10.18
packaging                 25.0
pandas                    2.2.0
parso                     0.8.4
pexpect                   4.9.0
pillow                    10.2.0
pip                       24.0
platformdirs              4.3.8
prompt_toolkit            3.0.51
propcache                 0.3.1
protobuf                  6.31.1
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
py-cpuinfo                9.0.0
pycocotools               2.0.10
pycodestyle               2.11.1
pydantic                  1.10.7
pydantic_core             2.33.2
pyDeprecate               0.3.2
pydub                     0.25.1
pyflakes                  3.2.0
Pygments                  2.19.1
pynvml                    12.0.0
pyparsing                 3.2.3
python-dateutil           2.9.0.post0
python-multipart          0.0.20
pytorch-lightning         1.6.0
pytz                      2025.2
PyWavelets                1.8.0
PyYAML                    6.0.1
qudida                    0.0.4
qwen-vl-utils             0.0.11
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
rich                      14.0.0
rpds-py                   0.25.1
ruff                      0.11.13
safehttpx                 0.1.6
safetensors               0.5.3
scikit-image              0.20.0
scikit-learn              1.7.0
scipy                     1.15.3
semantic-version          2.10.0
sentencepiece             0.2.0
setuptools                78.1.1
shellingham               1.5.4
shortuuid                 1.0.13
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.7
stack-data                0.6.3
starlette                 0.46.2
supervision               0.25.1
svgwrite                  1.4.3
sympy                     1.14.0
tensorboard               2.19.0
tensorboard-data-server   0.7.2
termcolor                 3.1.0
threadpoolctl             3.6.0
tifffile                  2025.5.10
tiktoken                  0.9.0
timm                      0.6.13
tokenizers                0.13.3
tomli                     2.2.1
tomlkit                   0.13.3
torch                     2.2.1+cu118
torchmetrics              1.7.2
torchvision               0.17.1+cu118
tqdm                      4.67.1
traitlets                 5.14.3
transformers              4.29.2
triton                    2.2.0
typer                     0.16.0
typing_extensions         4.14.0
typing-inspection         0.4.1
tzdata                    2025.2
uc-micro-py               1.0.3
urllib3                   2.4.0
uvicorn                   0.34.3
wavedrom                  2.0.3.post3
wcwidth                   0.2.13
webdataset                0.2.86
websockets                15.0.1
Werkzeug                  3.1.3
wheel                     0.45.1
xformers                  0.0.25+cu118
yapf                      0.43.0
yarl                      1.20.0
zipp                      3.22.0
